requirements.txt
flask
flask-cors
apscheduler
requests
groq
sendgrid
python-dotenv
gunicorn
